Update README.md
This was my first lab in term 2 in my Programming class Object Oriented. The class, Medals, calculates and displays the total medals won by each country in an Olympic style event.
This is where I learned and use both the 1 and 2 dimensional arrays, and take their information and have them in a nested for loops. Calculating their information and output and display them in an olympic style event using the %15s and %8d.